+++ src/api/common.ts
@@ -1,0 +1,9 @@
+export const sendNotification = (content: string) => {
+    ipcRenderer.send('send-notification', content);
+};
+export const showOnTray = (content: string) => {
+    ipcRenderer.send('update-tray', content);
+};
+export const reloadWindow = () => {
+    ipcRenderer.send('reload-window');
+};
